requirement.txt
Django==4.1.7
djangorestframework==3.15.2
scikit-learn==1.6.1
numpy==2.2.2
pandas==2.2.3
